User roles and permission

+++ app/Http/Controllers/PostController.php
@@ -16,7 +16,8 @@ class PostController extends Controller
      */
     public function index()
     {
-        #sleep(3);
+        $this->authorize('viewAny', Post::class);
+
         $posts = PostResource::collection(Post::all());
 
         return inertia('Posts/Index', compact('posts'));
@@ -29,6 +30,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Post::class);
         return inertia('Posts/Create');
     }
 
@@ -40,6 +42,7 @@ class PostController extends Controller
      */
     public function store(StorePostRequest $request)
     {
+        $this->authorize('create', Post::class);
         Post::create($request->validated());
 
         return redirect()->route('posts.index')->with('message', 'Post created successfully');
@@ -64,8 +67,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        //
-        //$post = new PostResource($post);
+        $this->authorize('create', Post::class);
 
         return inertia('Posts/Edit', compact('post'));
     }
@@ -79,6 +81,7 @@ class PostController extends Controller
      */
     public function update(StorePostRequest $request, Post $post)
     {
+        $this->authorize('create', Post::class);
         $post->update($request->validated());
 
         return redirect()->route('posts.index')->with('message', 'Post updated successfully');
@@ -93,6 +96,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('create', Post::class);
         $post->delete();
         return redirect()->route('posts.index')->with('message', 'Post deleted successfully');
     }
